% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotarrow-package.R
\docType{package}
\name{hotarrow-package}
\alias{hotarrow}
\alias{hotarrow-package}
\title{hotarrow: Right Arrow Add to Addin to Use Shortcut}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
