% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_columns.R
\name{insert_columes_arrow}
\alias{insert_columes_arrow}
\title{Insert %$%.}
\usage{
insert_columes_arrow()
}
\description{
Call this function as an addin to insert \code{\code{\%$\%}} at the cursor position.
}
