% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeallot.R
\name{insert_left_arrow_zeallot}
\alias{insert_left_arrow_zeallot}
\title{Insert %<-%.}
\usage{
insert_left_arrow_zeallot()
}
\description{
Call this function as an addin to insert \code{\code{\%<-\%}} at the cursor position.
}
