% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_arrow.R, R/tee.R
\name{insert_right_arrow}
\alias{insert_right_arrow}
\title{Insert ->.}
\usage{
insert_right_arrow()

insert_right_arrow()
}
\description{
Call this function as an addin to insert \code{ -> } at the cursor position.
}
