% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeallot.R
\name{insert_right_arrow_zeallot}
\alias{insert_right_arrow_zeallot}
\title{Insert %->%.}
\usage{
insert_right_arrow_zeallot()
}
\description{
Call this function as an addin to insert \code{\code{\%->\%}} at the cursor position.
}
